1. index.php - Menambah kondisi jika belum ada data sama sekali; tamb…

+++ index.php
@@ -1,6 +1,8 @@
 <?php include 'koneksi.php'; ?>
 
-<a href="tambah_mahasiswa.php">Tambah Data</a>
+<h1>Data Mahasiswa</h1>
+
+<p><a href="tambah_mahasiswa.php">Tambah Data</a></p>
 <table border="1">
     <tr>
         <th>NIM</th>
@@ -11,17 +13,22 @@
     </tr>
     <?php
     $data = mysqli_query($koneksi, "SELECT * FROM mahasiswa");
-    while ($row = mysqli_fetch_array($data)) {
-        echo "<tr>
-            <td>" . $row['nim'] . "</td>
-            <td>" . $row['nama'] . "</td>
-            <td>" . $row['jenis_kelamin'] . "</td>
-            <td>" . $row['alamat'] . "</td>
-            <td>
-                <a href='edit.php?id=" . $row['id'] . "'>Edit</a> | 
-                <a href='hapus.php?id=" . $row['id'] . "'>Hapus</a>
-            </td>
-        </tr>";
+    $row = mysqli_fetch_array($data);
+    if ($row) {
+        while ($row) {
+            echo "<tr>
+                <td>" . $row['nim'] . "</td>
+                <td>" . $row['nama'] . "</td>
+                <td>" . $row['jenis_kelamin'] . "</td>
+                <td>" . $row['alamat'] . "</td>
+                <td>
+                    <a href='edit.php?id=" . $row['id'] . "'>Edit</a> | 
+                    <a href='hapus.php?id=" . $row['id'] . "'>Hapus</a>
+                </td>
+            </tr>";
+        }
+    } else {
+        echo "<tr><td colspan='5'>Belum ada data.</td></tr>";
     }
     ?>
 </table>
